Product.Quantity from decimal type to double type (nullable)

+++ RevisoSharp/RevisoItems/Product.cs
@@ -158,7 +158,7 @@ namespace RevisoSharp.RevisoItems
         /// </summary>
         [JsonPropertyName("quantity")]
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public decimal? Quantity { get; set; }
+        public double? Quantity { get; set; }
 
         /// <summary>
         ///
